refactor post api endpoint
